% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_cypher_node.R
\name{df_to_cypher_node}
\alias{df_to_cypher_node}
\title{Convert a data frame to Cypher node creation statements}
\usage{
df_to_cypher_node(df, label = "Node", id_column = NULL, use_merge = FALSE)
}
\arguments{
\item{df}{A data frame with node data}

\item{label}{The node label}

\item{id_column}{Column to use as the node identifier}

\item{use_merge}{Logical, whether to use MERGE instead of CREATE}
}
\value{
A character vector of Cypher statements
}
\description{
Convert a data frame to Cypher node creation statements
}
