% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_column_names.R
\name{normalize_column_names}
\alias{normalize_column_names}
\title{Normalize column names to snake_case}
\usage{
normalize_column_names(df)
}
\arguments{
\item{df}{A data.frame or tibble}
}
\value{
Dataframe with normalized column names
}
\description{
Normalize column names to snake_case
}
