% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_cypher.R
\name{preview_cypher}
\alias{preview_cypher}
\title{Preview Cypher Statements in Console}
\usage{
preview_cypher(cypher_lines)
}
\arguments{
\item{cypher_lines}{A character vector of Cypher statements}
}
\value{
No return value. Prints to console.
}
\description{
Preview Cypher Statements in Console
}
\examples{
preview_cypher(c("CREATE (:Person {name: 'Alice'})"))
}
