% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cypher_file.R
\name{write_cypher_file}
\alias{write_cypher_file}
\title{Write Cypher Statements to a File}
\usage{
write_cypher_file(cypher_lines, file, append = FALSE)
}
\arguments{
\item{cypher_lines}{A character vector of Cypher statements}

\item{file}{Path to output file (e.g., "output.cypher")}

\item{append}{Logical. Should output be appended to the file?}
}
\value{
The path to the written file (invisible)
}
\description{
Write Cypher Statements to a File
}
\examples{
write_cypher_file(c("CREATE (:Person {name: 'Alice'})"), "example.cypher")
}
